chore: update and pin actions in GH workflows
Updating these since some were a bit old. Also, pinning to the hash is a
good security practice. The pinning was automated with pinact (running
`pinact run --update`).

This now allows configuring in the repo the blocking of actions that are
not pinned (see [GitHub Actions policy now supports blocking and SHA
pinning actions]).

Closes: #1340, #1341, #1342, #1353

[GitHub Actions policy now supports blocking and SHA pinning actions]: https://github.blog/changelog/2025-08-15-github-actions-policy-now-supports-blocking-and-sha-pinning-actions/